requirements.txt
##############################################################################
# Core dependencies
##############################################################################
openai>=1.30.0
pyyaml>=6.0

##############################################################################
# Optional: Anthropic provider
# Uncomment: pip install anthropic>=0.25.0
##############################################################################
# anthropic>=0.25.0

##############################################################################
# Optional: load .env files automatically
# Uncomment: pip install python-dotenv>=1.0.0
##############################################################################
# python-dotenv>=1.0.0

##############################################################################
# Development / testing
##############################################################################
pytest>=8.0.0
